valk: Extract courier library
